requirements.txt
numba==0.43.1

# python 3.7
# llvmlite==0.32.1 			
llvmlite==0.31.0			

filterpy==1.4.5
scikit-learn==0.19.2
matplotlib==2.2.3
pillow==6.2.2
opencv-python==3.4.3.18
glob2==0.6

open3d==0.10.0.1
alfred-py==2.7.5